requirements.txt
gradio 
numpy 
Pillow 
scipy 
scikit-image 
torch 
opencv-python